fix: error behavior of clzw 0b1xxx
